Moved all header only code into a header directory. Using an index.js to export from the new directory.
