+++ imports/ui/components/header/worker-header.jsx
@@ -1,13 +1,15 @@
 import React from "react";
-import { If, Then, Else } from "../if-else.jsx";
-import LangSelector from "../components/lang-selector.jsx";
+
+import { Meteor } from "meteor/meteor";
 import i18n from "meteor/universe:i18n";
-import { urlGenerators } from "../../startup/client/router.jsx";
 import { withTracker } from "meteor/react-meteor-data";
+
+import { If, Then, Else } from "/imports/ui/if-else.jsx";
+import LangSelector from "./lang-selector.jsx";
 
 const T = i18n.createComponent();
 
-class EmployerHeader extends React.Component {
+class WorkerHeader extends React.Component {
 	render() {
 		return (
 			<div className="top-nav">
@@ -61,24 +63,30 @@ class EmployerHeader extends React.Component {
 									</If>
 								</li>
 								<li>
-									<a href="" className="link-kumya ">
-										<span>My Company</span>
+									<a
+										href="/companies"
+										className="link-kumya "
+									>
+										<span>
+											<T>common.header.companies</T>
+										</span>
+									</a>
+								</li>
+								<li>
+									<a href="/jobs" className="link-kumya">
+										<span>
+											<T>common.header.jobs</T>
+										</span>
 									</a>
 								</li>
 								<li>
 									<a
-										href="/post-a-job"
+										href="/worker-resources"
 										className="link-kumya"
 									>
-										<span>Post a Job</span>
-									</a>
-								</li>
-								<li>
-									<a
-										href="/employer-resources"
-										className="link-kumya"
-									>
-										<span>Resources</span>
+										<span>
+											<T>common.header.resources</T>
+										</span>
 									</a>
 								</li>
 							</ul>
@@ -100,7 +108,7 @@ class EmployerHeader extends React.Component {
 												<li className="tr">
 													<a
 														href="/my-account"
-														className="link-kumya"
+														className="btn navbar-btn margin-right btn-green hvr-icon-forward"
 													>
 														{i18n.__(
 															"common.header.myaccount"
@@ -110,7 +118,7 @@ class EmployerHeader extends React.Component {
 												<li className="tr">
 													<a
 														onClick={Meteor.logout}
-														className="link-kumya"
+														className="navbar-link margin-right"
 														style={{
 															cursor: "pointer",
 														}}
@@ -150,13 +158,23 @@ class EmployerHeader extends React.Component {
 								<li className="dropdown">
 									<LangSelector />
 								</li>
+								<li>
+									<a
+										href="/foremployers"
+										className="link-kumya"
+									>
+										<span>
+											<T>common.header.for_employers</T>
+										</span>
+									</a>
+								</li>
 								<br />
 								<If cond={this.props.isLoggedIn}>
 									<Then>
 										<li>
 											<a
 												onClick={Meteor.logout}
-												className="toggle-only-display link-kumya"
+												className="toggle-only-display navbar-link margin-right"
 												style={{ cursor: "pointer" }}
 											>
 												<T>common.header.logout</T>
@@ -177,4 +195,4 @@ class EmployerHeader extends React.Component {
 
 export default withTracker(() => ({
 	isLoggedIn: Meteor.userId() !== null,
-}))(EmployerHeader);
+}))(WorkerHeader);
