+++ imports/ui/pages/showjobs.jsx
@@ -5,7 +5,7 @@ import { Query } from "react-apollo";
 import i18n from "meteor/universe:i18n";
 
 import ShowJobComponent from "/imports/ui/components/showJobComponent.jsx";
-import Header from "/imports/ui/components/header.jsx";
+import Header from "/imports/ui/components/header";
 
 const T = i18n.createComponent();
